backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pgvector==0.3.6
python-multipart==0.0.9
jinja2==3.1.4
httpx==0.27.2
itsdangerous==2.2.0
pydantic==2.10.2
pydantic-settings==2.6.1
cryptography==43.0.3
numpy==2.1.3
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
